Fix up for decoupled BLS
